+++ src/lib.rs
@@ -6,6 +6,7 @@
 
 pub mod accents;
 pub mod builders;
+pub mod feature_classes;
 pub mod features;
 pub mod phoneme;
 pub mod stress;